Seguimos con el curso, avanzando poco a poco

+++ 01_ring/ft_strlcat.c
@@ -17,7 +17,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t size)
 	size_t	i;
 	size_t	j;
 
-	if (size <= ft_strlen(dst))
+	if (size <= (size_t)ft_strlen(dst))
 		return (size + ft_strlen(src));
 	i = ft_strlen(dst);
 	j = 0;
